+++ src/user/user.service.ts
@@ -2,7 +2,7 @@ import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { IFirebaseUser } from '../firebase/firebase-user.interface';
 import { UserRepository } from './user.repository';
 import { InjectRepository } from '@nestjs/typeorm';
-import { formatUserObject, getUserData } from '../utils/constants';
+import { formatUserObject, getCrispUserData } from '../utils/serialize';
 import { GetUserDto } from './dtos/get-user.dto';
 import { CreateUserDto } from './dtos/create-user.dto';
 import { UserEntity } from '../entities/user.entity';
@@ -10,7 +10,7 @@ import { PartnerAccessService } from '../partner-access/partner-access.service';
 import { PartnerAccessEntity } from '../entities/partner-access.entity';
 import { PartnerEntity } from '../entities/partner.entity';
 import { PartnerRepository } from '../partner/partner.repository';
-import { addNewPeopleProfile, savePeopleData } from 'src/api/crisp/api-crisp';
+import { addCrispProfile, updateCrispProfile } from 'src/api/crisp/api-crisp';
 
 @Injectable()
 export class UserService {
@@ -47,24 +47,24 @@ export class UserService {
             createUserResponse.id,
           );
 
-        const partnerDetails = await this.partnerRepository.findOne({
+        const getPartnerResponse = await this.partnerRepository.findOne({
           id: updatePartnerAccessResponse.partnerId,
         });
 
         if (!!updatePartnerAccessResponse.featureLiveChat) {
           const {
             data: { data },
-          } = await addNewPeopleProfile({
+          } = await addCrispProfile({
             email: createUserResponse.email,
             person: { nickname: createUserResponse.name },
           });
 
-          const userData = getUserData(
+          const userData = getCrispUserData(
             createUserResponse,
-            partnerDetails,
+            getPartnerResponse,
             updatePartnerAccessResponse,
           );
-          await savePeopleData({ ...userData }, data?.people_id);
+          await updateCrispProfile({ ...userData }, data?.people_id);
         }
 
         delete updatePartnerAccessResponse.partnerAdmin;
@@ -72,7 +72,7 @@ export class UserService {
         return {
           user: createUserResponse,
           partnerAccess: updatePartnerAccessResponse,
-          partner: partnerDetails,
+          partner: getPartnerResponse,
         };
       }
